backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
django-filter
python-dotenv
psycopg2-binary
pgvector
django-htmx
gunicorn
uvicorn
Pillow
pillow-heif
numpy
sentence-transformers
torch
wcwidth==0.2.14
insightface>=0.7.3
onnxruntime-gpu
opencv-python-headless
